requirements.txt
playwright>=1.40.0
flask>=2.0.0
openai>=1.0.0
python-dotenv>=0.19.0 